+++ CHANGELOG.md
@@ -22,10 +22,11 @@
 
 ### Fixed
 
-- Fix filter component initial active state
+- Fix Filter component's initial active state
 - Fix translating `x` and `y` properties simultaneously in Parallax component
 - Class utils can handle SVGs in IE 11 now
 - Fix drag closes Lightbox
+- Fix Tooltip component (touch device)
 
 ## 3.0.3 (January 29, 2019)